[fix] 修復 Swift 編譯錯誤與協議一致性問題
- 新增 ObservableObject 一致性至 AdaptiveFrameRateController 和 MicroGestureDetector
- 修復 CalibrationStatus 重複宣告，重新命名為 CalibrationWizardStatus
- 實作 Hashable/Equatable 一致性至手勢相關類型
- 新增完整 Codable 支援至所有資料結構
- 修復 InstrumentMode 缺少的屬性和方法
- 解決 BrowseView 和 MicroGestureDetector 的類型錯誤
- 改善資料持久化和序列化支援

+++ GazeTurn/View/DashboardView.swift
@@ -330,7 +330,7 @@ struct DashboardView: View {
 
                 HealthIndicator(
                     title: "處理負載",
-                    value: frameRateController.performanceMetrics.averageFrameTime / 16.67,
+                    value: Float(frameRateController.performanceMetrics.averageFrameTime / 16.67),
                     format: "%.1fx",
                     goodThreshold: 1.5,
                     warningThreshold: 2.0
